Added SQL database scripts and Password query
The MySQL folder contains the scripts to create the database and to fill it with some test data.

server.js was updated to send a query to the database to check if the password is the same stored for the provided email
(right now it is named username, but it will in the future be changed to email).